Enterprise License Support , RocksDB is  default storage engine so all references to MMfiles have been removed ,

+++ arangosh/Benchmark/testcases/RandomShapesTest.h
@@ -45,7 +45,7 @@ namespace arangodb::arangobench {
 
     void tearDown() override {}
 
-    void buildRequest(int threadNumber, size_t threadCounter,
+    void buildRequest(size_t threadNumber, size_t threadCounter,
                       size_t globalCounter, BenchmarkOperation::RequestData& requestData) const override {
       size_t keyId = static_cast<size_t>(globalCounter / 3);
       std::string const key = "testkey" + StringUtils::itoa(keyId);
