+++ arangosh/Benchmark/testcases/StreamCursorTest.h
@@ -43,7 +43,7 @@ namespace arangodb::arangobench {
 
     void tearDown() override {}
 
-    void buildRequest(int threadNumber, size_t threadCounter,
+    void buildRequest(size_t threadNumber, size_t threadCounter,
                       size_t globalCounter, BenchmarkOperation::RequestData& requestData) const override {
       requestData.url = "/_api/cursor";
       requestData.type = rest::RequestType::POST;
